requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
werkzeug==3.0.6
Flask-Cors==4.0.0
gunicorn==21.2.0
